Move prepareNextActions from utils to sagas
